+++ include/kvm/irq.h
@@ -44,6 +44,10 @@ int irq__common_add_irqfd(struct kvm *kvm, unsigned int gsi, int trigger_fd,
 			   int resample_fd);
 void irq__common_del_irqfd(struct kvm *kvm, unsigned int gsi, int trigger_fd);
 
+#ifdef CONFIG_X86
+bool irqchip_split(struct kvm *kvm);
+#endif
+
 #ifndef irq__add_irqfd
 #define irq__add_irqfd irq__common_add_irqfd
 #endif
